requirements.txt
torch==2.4.0
shortuuid
fastapi
uvicorn
transformers==4.45.1
pillow==10.2.0
matplotlib==3.9.2
seaborn==0.13.2
openai==1.51.2
httpx==0.23.3
ray==2.37.0
pyzmq==26.2.0